Call sortable's `refresh` method after content changes. See #2.
This fixes an issue where the "ui-sortable-handle" class would not be
added to newly-inserted list items.

To make this work, I have to add a second array observer (in addition to
the one added by `Ember.CollectionView`) to the content array, that is
guaranteed to fire only after `didInsertElement` is triggered.